fix(inclusionCalculator): Move MultiMatcher to files package
